Address review: safer topics backoff, cache cleanups, allowlist module
- task_traces: use a None-safe sort key so a missing task id can't raise
  TypeError and drop all config/plugin events for the Connect URL.
- topics fetch: back off only successful fetches and permanently-unsupported
  (404/405) endpoints; retry transient failures on the next run, matching
  _collect_plugins. Only rewrite the topics data cache when it actually changed.
- cache: delegate get_items_to_fetch/mark_items_fetched/get_events_to_send to
  the new get_cached_json/set_cached_json helpers instead of duplicating the
  read/write-with-try/except blocks.
- Move ALLOWED_CONFIG_KEYS/SENSITIVE_KEY_SUBSTRINGS/is_allowed_config_key into
  connector_config_keys.py so connectors.py stays focused on collection.
- Add a .changed changelog fragment for the denylist->allowlist redaction flip.
- Add tests for transient-vs-unsupported topics backoff and None task ids.
