+++ kafka_consumer/tests/test_connectors.py
@@ -273,6 +273,64 @@ def test_config_event_survives_topics_fetch_failure(run_connect_check, aggregato
     assert all(event['topics'] == [] for event in events)
 
 
+class FakeHttpError(Exception):
+    """Stand-in for a requests HTTPError carrying an HTTP status code."""
+
+    def __init__(self, status_code):
+        super().__init__(f"HTTP {status_code}")
+        self.response = mock.MagicMock(status_code=status_code)
+
+
+def test_topics_transient_failure_retried_next_run(run_connect_check):
+    def get(url, **kwargs):
+        if url.endswith('/topics'):
+            raise ConnectionError("refused")
+        response = mock.MagicMock()
+        response.json.return_value = {} if 'connector-plugins' in url else SAMPLE_CONNECTORS_RESPONSE
+        return response
+
+    _, http = run_connect_check(get_side_effect=get, runs=2)
+
+    topic_fetches = [call for call in http.get.call_args_list if call.args[0].endswith('/topics')]
+    assert len(topic_fetches) == 4, "a transient topics failure should be retried on the next run, not backed off"
+
+
+def test_topics_unsupported_endpoint_backed_off(run_connect_check):
+    def get(url, **kwargs):
+        if url.endswith('/topics'):
+            raise FakeHttpError(404)
+        response = mock.MagicMock()
+        response.json.return_value = {} if 'connector-plugins' in url else SAMPLE_CONNECTORS_RESPONSE
+        return response
+
+    _, http = run_connect_check(get_side_effect=get, runs=2)
+
+    topic_fetches = [call for call in http.get.call_args_list if call.args[0].endswith('/topics')]
+    assert len(topic_fetches) == 2, "an unsupported (404) topics endpoint should be backed off after one attempt"
+
+
+def test_task_traces_sort_handles_missing_task_id(run_connect_check, aggregator):
+    connectors = {
+        'my-sink': {
+            'info': {'type': 'sink', 'config': {'connector.class': 'io.confluent.SomeSink'}},
+            'status': {
+                'connector': {'state': 'RUNNING'},
+                'tasks': [
+                    {'id': None, 'state': 'FAILED', 'trace': 'trace-a'},
+                    {'id': None, 'state': 'FAILED', 'trace': 'trace-b'},
+                    {'id': 2, 'state': 'FAILED', 'trace': 'trace-c'},
+                ],
+            },
+        }
+    }
+    run_connect_check(connectors_response=connectors)
+
+    events = dsm_events(aggregator, 'connector')
+    assert len(events) == 1
+    # Mixed None/int task ids must not raise; the real integer id sorts ahead of the None ids.
+    assert [t['task_id'] for t in events[0]['task_traces']] == [2, None, None]
+
+
 def test_config_event_not_reemitted_when_unchanged(run_connect_check, aggregator):
     run_connect_check(connectors_response=SAMPLE_CONNECTORS_RESPONSE, runs=2)
